chore(pbf): move timeline stripe view to widget package
